+++ tests/testthat/test-model-sample.R
@@ -15,7 +15,7 @@ if (not_on_cran()) {
   # these are all valid for sample()
   ok_arg_values <- list(
     data = data_list,
-    output_dir = tempdir(),
+    dir = tempdir(),
     chains = 2,
     parallel_chains = 1,
     iter_warmup = 50,
@@ -39,7 +39,7 @@ if (not_on_cran()) {
   # using any one of these should cause sample() to error
   bad_arg_values <- list(
     data = "NOT_A_FILE",
-    output_dir = "NOT_A_DIRECTORY",
+    dir = "NOT_A_DIRECTORY",
     chains = -1,
     parallel_chains = -1,
     iter_warmup = -1,
@@ -62,7 +62,7 @@ if (not_on_cran()) {
 
   bad_arg_values_2 <- list(
     data = matrix(1:10),
-    output_dir = 1,
+    dir = 1,
     chains = "NOT_A_NUMBER",
     parallel_chains = "NOT_A_NUMBER",
     init = "NOT_A_FILE",